docs, contextlib, verbose, rm_tmp, RuntimeError raise when log lammps read fails, os.system(tail etc) replaced

+++ aelm.py
@@ -10,11 +10,18 @@
 # DOCS
 # ============================================================================
 
-"""Accelerated exploration of local minima."""
+"""Accelerated exploration of local minima.
+
+This is a fast and effective method to find relevant minima in which an energy
+bias function is used to overcome the energetic barriers followed by a local
+minimization.
+"""
 
 # ============================================================================
 # IMPORTS
 # ============================================================================
+
+import contextlib
 import os
 import subprocess
 
@@ -34,6 +41,8 @@ def aelm(
     lmp_data="in.frame",
     lmp_min="dump.minimization.lammpstrj",
     lmp_flags=None,
+    verbose=True,
+    rm_tmp=True,
 ):
     """Accelerated exploration of local minima minimization.
 
@@ -57,15 +66,22 @@ def aelm(
     lmp_in : str, default="in.minimization"
         the input file for LAMMPS
 
+    lmp_data : str, default="in.frame"
+        the file with the frame to be minimized by LAMMPS
+
     lmp_min : str, default="dump.minimization.lammpstrj"
         the dump file where the local minimization is written
-
-    lmp_in : str, default="in.minimization"
-        the file with the frame to be minimized by LAMMPS
 
     lmp_flags : dict, default=None
         command-line options for LAMMPS, where the key is the flag and the
         value the option.
+
+    verbose : bool, default=True
+        print on the screen each of the values obtained for the performed
+        minimizations
+
+    rm_tmp : bool, default=True
+        remove tmp files
 
     Returns
     -------
@@ -74,19 +90,38 @@ def aelm(
         final energies of each frame. It also generates a lammpstrj file with
         all the minimum energy frames in the same order as the initial biased
         xyz file.
+
+    Raises
+    ------
+    RuntimeError
+        A problem has occurred, it may be because the LAMMPS simulation has
+        not finished correctly or the log file does not correspond to a
+        minimization.
     """
+    if verbose:
+        k = 0
+        print(
+            "# minimization number, initial energy, next_to_last energy, "
+            "final energy"
+        )
+
     initial, next_to_last, final = [], [], []
-    with exma.io.reader.XYZ(biased_traj, ftype="xyz") as bias:
+    with contextlib.ExitStack() as stack:
+        bias_traj = stack.enter_context(
+            exma.io.reader.XYZ(biased_traj, ftype="xyz")
+        )
+        min_traj = stack.enter_context(
+            exma.io.writer.LAMMPS(minimizations_output)
+        )
         try:
             while True:
-                # read and xyz frame
-                frame = bias.read_frame()
+                frame = bias_traj.read_frame()
+
                 frame.box = cell_info["box"]
                 frame.idx = np.arange(1, frame.natoms + 1)
                 frame.types = [cell_info["type"][t] for t in frame.types]
                 frame.q = np.zeros(frame.natoms, dtype=np.float32)
 
-                # write the frame to a configurations input file for lammps
                 exma.io.writer.in_lammps(lmp_data, frame)
 
                 # run lammps minimization with the corresponding flags
@@ -102,33 +137,51 @@ def aelm(
                 log = lmp_run.stdout.split("\n")
 
                 # get the energies and save for the pd.DataFrame
-                for line, next_line in zip(log, log[1:] + [log[0]]):
-                    if line.strip().startswith("Energy"):
-                        e1, e2, e3 = next_line.strip().split()
-                initial.append(e1)
-                next_to_last.append(e2)
-                final.append(e3)
+                try:
+                    for line, next_line in zip(log, log[1:] + [log[0]]):
+                        if line.strip().startswith("Energy initial"):
+                            e1, e2, e3 = next_line.strip().split()
+                    initial.append(e1)
+                    next_to_last.append(e2)
+                    final.append(e3)
+                except UnboundLocalError:
+                    raise RuntimeError(
+                        "A problem has occurred, it may be because the LAMMPS "
+                        "simulation has not finished correctly or the log "
+                        "file does not correspond to a minimization."
+                    )
 
                 # accumulate the minimum energy frames
-                # idk how to generalize this line if i dont want to use
-                # tail or os.system
-                os.system(
-                    f"tail -{frame.natoms + 9} {lmp_min}"
-                    f" >> {minimizations_output}"
+                with exma.io.reader.LAMMPS(lmp_min) as tmp:
+                    min_frame = tmp.read_frame()
+                    try:
+                        while True:
+                            tmp_frame = tmp.read_frame()
+                            min_frame = tmp_frame
+                    except EOFError:
+                        ...
+                min_traj.write_frame(
+                    min_frame._sort_frame()
+                    if not min_frame._is_sorted()
+                    else min_frame
                 )
 
-                # remove tmp files
-                for file in [lmp_data, lmp_min, "log.cite", "log.lammps"]:
-                    os.remove(file)
+                if rm_tmp:
+                    for file in (lmp_data, lmp_min, "log.cite", "log.lammps"):
+                        os.remove(file)
+
+                if verbose:
+                    print(k, e1, e2, e3)
+                    k += 1
 
         except EOFError:
             ...
 
         finally:
-            return pd.DataFrame(
-                {
-                    "initial": np.asarray(initial, dtype=np.float32),
-                    "next_to_last": np.asarray(next_to_last, dtype=np.float32),
-                    "final": np.asarray(final, dtype=np.float32),
-                }
-            )
+            initial = np.asarray(initial, dtype=np.float32)
+            next_to_last = np.asarray(next_to_last, dtype=np.float32)
+            final = np.asarray(final, dtype=np.float32)
+
+    return pd.DataFrame(
+        {"initial": initial, "next_to_last": next_to_last, "final": final}
+    )
